requirements.txt
aiohttp>=3.8.0
psycopg2>=2.9.0
requests>=2.26.0
python-dateutil>=2.8.2
scipy>=1.8.0
sphinx>=7.0.0
sphinx-press-theme>=0.8.0
PyQt6>=6.0.0